PetAdoption/requirements.txt
asgiref==3.8.1
Django==5.1.5
pillow==11.1.0
sqlparse==0.5.3
tzdata==2024.2
djangorestframework==3.15.2
django-filter==23.5
django-cors-headers
